apalis-imx8: board: remove function pmic_match_otp
PMICs are usually leaving the factory of NXP in a fused state. We do
order them unfused and fuse them in our production.

The function pmic_match_otp is used to check if an ID inside those fuses
match and hence it checks if the PMIC is properly fused. We do not fuse
any ID which makes the whole check of pmic_match_otp useless.

Remove the function

Related-to: ELB-4478
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>