mx8qx_colibri: improve setup debug-uart
So far, if one wanted to compile the SCFW for Colibri iMX8X to use the
Colibri UART_A (standard debug uart) using compile-option U=6, it was
dependending on a change in platform/devices/MX8QX/MX8QX.h.
This file however is overwritten in case of a SCFW update.

Change setting up the debug-uart like NXP does it.

Related-to: ELB-4574
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>